Added InputGroupStyle test
