+++ cypress/cypress/component/Input/BasicInputField.cy.tsx
@@ -346,37 +346,4 @@ describe("Input.cy.tsx", () => {
 
     cy.get(`label[for=${name}]`).parent().find("svg[data-icon=pencil]");
   });
-
-  it("contains pencil icon and inputGroup has correct style", () => {
-    const name = faker.random.alpha(10);
-    const schema = yup.object().shape({
-      [name]: yup.date(),
-    });
-    const inputGroupStyle = "background-color: black;";
-    const defaultStyle = "flex-wrap: nowrap; ";
-    const style = defaultStyle.concat(inputGroupStyle);
-
-    cy.mount(
-      <Form
-        onSubmit={() => {
-          // Nothing to do
-        }}
-        resolver={yupResolver(schema)}
-      >
-        <Input
-          name={name}
-          label={name}
-          inputGroupStyle={{ backgroundColor: "black" }}
-          addonRight={
-            <InputGroupText>
-              <FontAwesomeIcon icon={faPencil} />
-            </InputGroupText>
-          }
-        />
-      </Form>,
-    );
-
-    cy.get(`label[for=${name}]`).parent().find("svg[data-icon=pencil]");
-    cy.get(".input-group").should("have.attr", "style", style);
-  });
 });
